requirements/base.txt
amqp==2.3.2
Babel==2.6.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.9
django-anymail==5.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-fieldsignals==0.3.4
django-filter==2.0.0
django-fsm==2.6.0
django-model-utils==3.1.2
django-sslserver==0.20
django-tables2==2.0.3
djangorestframework==3.9.0
EasyProcess==0.2.3
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
flower==0.9.2
idna==2.8
jdcal==1.4
jieba3k==0.35.1
kombu==4.2.2
lxml==4.3.0
Markdown==3.0.1
newspaper3k==0.2.8
nltk==3.4
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.5.12
pandas==0.23.4
Pillow==5.4.1
psycopg2-binary==2.7.5
python-dateutil==2.7.5
pytz==2018.7
PyVirtualDisplay==0.2.1
PyYAML==3.13
redis==3.0.1
requests==2.21.0
requests-file==1.4.3
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7.3
tablib==0.12.1
tinysegmenter==0.3
tldextract==2.2.0
tornado==5.1.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
